Refactor sql/connector=>sql/connection
Improves clarity and eases the next step of the plugin refactor
