+++ src/server/plugins/user/sql.js
@@ -2,7 +2,7 @@
 import { camelizeKeys, decamelizeKeys, decamelize } from 'humps';
 import { has } from 'lodash';
 import bcrypt from 'bcryptjs';
-import knex from '../../../server/sql/connector';
+import knex from '../../../server/sql/connection';
 
 // Actual query fetching and transformation in DB
 export default class User {
